# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: B2 GEO/SEO Analyzer API
  version: 0.1.0
  description: API inicial para projetos, análises, findings, scores e planos de ação.
servers:
  - url: https://api.example.com/v1
security:
  - bearerAuth: []
...
paths:
...
  /analysis-runs:
    post:
      tags: [Analysis]
      operationId: createAnalysisRun
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnalysisRunRequest'
      responses:
        '202':
          description: Análise enfileirada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisRun'
        '422':
          $ref: '#/components/responses/ValidationError'
  /analysis-runs/{analysisRunId}:
    get:
      tags: [Analysis]
      operationId: getAnalysisRun
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/AnalysisRunId'
      responses:
        '200':
          description: Estado da análise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisRun'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /analysis-runs/{analysisRunId}/findings:
    get:
      tags: [Analysis]
      operationId: listFindings
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/AnalysisRunId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, info]
        - name: scoreType
          in: query
          schema:
            type: string
            enum: [seo, geo, ai_visibility]
      responses:
        '200':
          description: Findings paginados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingList'
  /analysis-runs/{analysisRunId}/scores:
    get:
      tags: [Analysis]
      operationId: getScores
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/AnalysisRunId'
      responses:
        '200':
          description: Scores da análise
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Score'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId:
      name: X-Organization-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
...
    AnalysisRunId:
      name: analysisRunId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
    NotFound:
      description: Recurso não encontrado
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationError:
      description: Dados inválidos
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    CreateAnalysisRunRequest:
      type: object
      required: [projectId, type, source, modules]
      properties:
        projectId: { type: string, format: uuid }
        type: { type: string, enum: [page, domain, ai_visibility, integration] }
        url: { type: string, format: uri }
        source: { type: string, enum: [extension, dashboard, schedule, api] }
        modules:
          type: array
          minItems: 1
          items: { type: string }
        pageCaptureId: { type: [string, 'null'], format: uuid }
    AnalysisRun:
      type: object
      required: [id, projectId, type, status, progress, createdAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        type: { type: string }
        status:
          type: string
          enum: [queued, running, partial, completed, failed, cancelling, cancelled]
        progress: { type: integer, minimum: 0, maximum: 100 }
        coverage: { type: [number, 'null'], minimum: 0, maximum: 100 }
        confidence: { type: [number, 'null'], minimum: 0, maximum: 100 }
        createdAt: { type: string, format: date-time }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
    Evidence:
      type: object
      required: [type]
      properties:
        type: { type: string }
        excerpt: { type: [string, 'null'] }
        selector: { type: [string, 'null'] }
        attributeName: { type: [string, 'null'] }
        attributeValue: { type: [string, 'null'] }
        data: { type: object, additionalProperties: true }
    Finding:
      type: object
      required: [id, ruleKey, scoreType, category, status, severity, title]
      properties:
        id: { type: string, format: uuid }
        ruleKey: { type: string }
        scoreType: { type: string, enum: [seo, geo, ai_visibility] }
        category: { type: string }
        status: { type: string, enum: [pass, fail, warn, info, not_applicable, not_tested, unknown, error] }
        severity: { type: string, enum: [critical, high, medium, low, info] }
        title: { type: string }
        explanation: { type: string }
        impact: { type: string }
        remediation: { type: string }
        acceptanceCriteria: { type: string }
        priority: { type: string, enum: [P0, P1, P2, P3, P4] }
        effort: { type: integer, minimum: 1, maximum: 5 }
        confidence: { type: number, minimum: 0, maximum: 1 }
        estimatedScoreGain: { type: number, minimum: 0 }
        evidences:
          type: array
          items: { $ref: '#/components/schemas/Evidence' }
    FindingList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Finding' }
        nextCursor: { type: [string, 'null'] }
    Score:
      type: object
      required: [scoreType, rulesetVersion, score, coverage, confidence]
      properties:
        scoreType: { type: string, enum: [seo, geo, ai_visibility, discoverability] }
        rulesetVersion: { type: string }
        score: { type: number, minimum: 0, maximum: 100 }
        coverage: { type: number, minimum: 0, maximum: 100 }
        confidence: { type: number, minimum: 0, maximum: 100 }
        categories: { type: object, additionalProperties: true }
        gates: { type: array, items: { type: object, additionalProperties: true } }
...
    Problem:
      type: object
      required: [title, status, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        correlationId: { type: string }